Improved logging and status tracking
Added a logrus logger to the application struct which has a hook that
will automatically update the bottom info bar (with appropriate styling
for log level), and add the log message to the log view. The log view
can now be opened with Ctrl+L or with the `Logs` button in the bottom
right.